[fix] redis 설정 통합
