--- src/main/java/camp/woowak/lab/infra/config/RedissConfiguration.java
@@ -1,53 +1,0 @@
-package camp.woowak.lab.infra.config;
-
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.redisson.spring.data.connection.RedissonConnectionFactory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedissConfiguration {
-
-	private static final Logger log = LoggerFactory.getLogger(RedissConfiguration.class);
-	@Value("${spring.data.redis.host}")
-	private String redisHost;
-
-	@Value("${spring.data.redis.port}")
-	private int redisPort;
-
-	private static final String REDISSON_HOST_PREFIX = "redis://";
-
-	@Bean
-	public RedissonClient redissonClient() {
-		log.info("Creating redisson config client start");
-		Config config = new Config();
-		config.useSingleServer().setAddress(REDISSON_HOST_PREFIX + redisHost + ":" + redisPort);
-		RedissonClient redissonClient = Redisson.create(config);
-		log.info("Creating redisson config client end");
-		return redissonClient;
-	}
-
-	@Bean
-	public RedisConnectionFactory redisConnectionFactory(RedissonClient redissonClient) {
-		return new RedissonConnectionFactory(redissonClient);
-	}
-
-	@Bean
-	public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-		RedisTemplate<String, Object> template = new RedisTemplate<>();
-		template.setConnectionFactory(redisConnectionFactory);
-
-		template.setKeySerializer(new StringRedisSerializer());
-		template.setValueSerializer(new GenericJackson2JsonRedisSerializer());
-		return template;
-	}
-}
